requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
torchinfo
calflops
einops
Pillow
tqdm
